Move response converter function selection into its own class; This makes it easier to test its behaviour
